LayoutTests/svg/W3C-SVG-1.1/interact-events-01-b-expected.txt

KCanvasContainer at (0,0) size 479x359
  KCanvasContainer at (77,44) size 331x214 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
    KCanvasContainer at (77,44) size 331x214 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
      KCanvasContainer at (77,44) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(77.500000,45.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasItem {rect} at (77,44) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
      KCanvasContainer at (202,44) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(202.500000,45.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasItem {rect} at (202,44) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
      KCanvasContainer at (327,44) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(327.500000,45.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasContainer at (327,44) size 81x81 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
          KCanvasItem {rect} at (327,44) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
      KCanvasContainer at (77,177) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(77.500000,178.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasContainer at (77,177) size 81x81
          KCanvasItem {rect} at (77,177) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
      KCanvasContainer at (202,177) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(202.500000,178.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasContainer at (202,177) size 81x81 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
          KCanvasItem {rect} at (202,177) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
      KCanvasContainer at (327,177) size 81x81 [transform={m=((1.000000,0.000000)(0.000000,1.000000)) t=(327.500000,178.000000)}] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}]
        KCanvasItem {rect} at (327,177) size 81x81 [HIDDEN] [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M0.00,0.00L80.00,0.00L80.00,80.00L0.00,80.00"]
  KCanvasItem {rect} at (0,0) size 479x359 [stroke={server=UNIMPLEMENTED}] [fill={server=UNIMPLEMENTED}] [data="M1.00,1.00L479.00,1.00L479.00,359.00L1.00,359.00"]
